requirements.txt
# Core Framework
flask==2.2.5
flask-cors==4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
aiohttp==3.9.3

# LangChain & AI
langchain==0.3.23
langchain-core==0.3.51
langchain-openai>=0.0.5
langgraph>=0.0.15
html2text>=2020.1.16

# OpenAI
openai>=1.3.0 